Initial work for multiple sources
