+++ pkg/ipv6/ipify.go
@@ -1,12 +1,10 @@
 package ipv6
 
 import (
+	log "github.com/juusujanar/cloudflare-ddns/pkg/logging"
 	"io/ioutil"
 	"net"
 	"net/http"
-	"strings"
-
-	log "github.com/juusujanar/cloudflare-ddns/pkg/logging"
 )
 
 // GetIPv6 gets IP from api6.ipify.org, which gives out only IPv6
@@ -24,14 +22,10 @@ func GetIPv6() (string, bool) {
 		log.Error("Could not get IPv6 address.")
 		return "", false
 	} else {
-		defer resp.Body.Close()
+		defer func() {
+			_ = resp.Body.Close()
+		}()
 		body, _ := ioutil.ReadAll(resp.Body)
 		return string(body), ValidateIPv6(string(body))
 	}
-}
-
-// ValidateIPv6 ensures that we got IPv6 address instead of IPv4
-func ValidateIPv6(str string) bool {
-	ip := net.ParseIP(str)
-	return ip != nil && strings.Contains(str, ":")
 }
